Added python script to generate grid topologies for simgrid
This tool is used as follows: python script_name.py number_rows number_columns
path_to_creation_hostfile path_to_creation_platform.
For convenience as this topology will be used only with the ring and grid algorithm
we implemented, only the routes that we use are declared, otherwise this can lead to
the explosion of the size of the generated file for a high number of processors
or to too many calculations for the routes.